new omponent created, not functional
