+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"C:\\Users\\daver\\OneDrive\\Documents\\sites\\juno\\week8\\project-6\\project-6\\src\\index.js":"1","C:\\Users\\daver\\OneDrive\\Documents\\sites\\juno\\week8\\project-6\\project-6\\src\\Search.js":"2","C:\\Users\\daver\\OneDrive\\Documents\\sites\\juno\\week8\\project-6\\project-6\\src\\Haiku.js":"3"},{"size":517,"mtime":1607278452365,"results":"4","hashOfConfig":"5"},{"size":2481,"mtime":1607384679756,"results":"6","hashOfConfig":"5"},{"size":7064,"mtime":1607384679754,"results":"7","hashOfConfig":"5"},{"filePath":"8","messages":"9","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"10"},"rrs5r5",{"filePath":"11","messages":"12","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":"13"},{"filePath":"14","messages":"15","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"C:\\Users\\daver\\OneDrive\\Documents\\sites\\juno\\week8\\project-6\\project-6\\src\\index.js",[],["16","17"],"C:\\Users\\daver\\OneDrive\\Documents\\sites\\juno\\week8\\project-6\\project-6\\src\\Search.js",["18"],"import { Component } from \"react\";\r\nimport axios from \"axios\";\r\nclass Search extends Component {\r\n  constructor() {\r\n    super();\r\n    this.state = {\r\n      suggestions: \"\",\r\n    };\r\n  }\r\n  //this function gets suggested words for autocomplete from api\r\n  getSuggestions = (word) => {\r\n    axios({\r\n      url: \"https://api.datamuse.com/sug\",\r\n      responseType: \"json\",\r\n      method: \"GET\",\r\n      params: {\r\n        s: word,\r\n      },\r\n    }).then(({ data }) => {\r\n      //get only first three suggestions\r\n      const suggestions = data.filter((item, index) => {\r\n        if (index < 3) {\r\n          return item;\r\n        }\r\n      });\r\n      //set the suggestions into state\r\n      this.setState({\r\n        suggestions: suggestions,\r\n      });\r\n    });\r\n  };\r\n  //when suggested word is clicked, this function is called. It gets the syllables number of the clicked word\r\n  getWordInfo = (e) => {\r\n    const word = e.target.dataset.name;\r\n    axios({\r\n      url: \"https://api.datamuse.com/words\",\r\n      responseType: \"json\",\r\n      method: \"GET\",\r\n      params: {\r\n        //this param tells api to pull the words that have similar prononcuation to the query. the first result will be the word itself\r\n        sp: word,\r\n        //this param tells api to include the number of syllables in response\r\n        md: \"s\",\r\n      },\r\n    }).then(({ data }) => {\r\n      console.log(data[0]);\r\n      //this function updates the state in app.js with the word that user clicked and the number of syllables in that word\r\n      this.props.updateSearchQuery(data[0]);\r\n    });\r\n  };\r\n  render() {\r\n    return (\r\n      <section className=\"Search\">\r\n        <form>\r\n          <label className=\"srOnly\" htmlFor=\"search\">\r\n            Input your word\r\n          </label>\r\n          <input\r\n            type=\"text\"\r\n            id=\"search\"\r\n            onChange={(e) => {\r\n              this.getSuggestions(e.target.value);\r\n            }}\r\n          />\r\n        </form>\r\n        <ul>\r\n          {/* displays suggested words (if there are any) as lis */}\r\n          {this.state.suggestions &&\r\n            this.state.suggestions.map((word) => {\r\n              return (\r\n                <li\r\n                  key={word.score}\r\n                  onClick={this.getWordInfo}\r\n                  data-name={word.word}\r\n                >\r\n                  {word.word}\r\n                </li>\r\n              );\r\n            })}\r\n        </ul>\r\n      </section>\r\n    );\r\n  }\r\n}\r\nexport default Search;\r\n","C:\\Users\\daver\\OneDrive\\Documents\\sites\\juno\\week8\\project-6\\project-6\\src\\Haiku.js",[],{"ruleId":"19","replacedBy":"20"},{"ruleId":"21","replacedBy":"22"},{"ruleId":"23","severity":1,"message":"24","line":21,"column":53,"nodeType":"25","messageId":"26","endLine":21,"endColumn":55},"no-native-reassign",["27"],"no-negated-in-lhs",["28"],"array-callback-return","Array.prototype.filter() expects a value to be returned at the end of arrow function.","ArrowFunctionExpression","expectedAtEnd","no-global-assign","no-unsafe-negation"]
+[{"D:\\sites\\juno\\bootcamp\\projects\\project-6\\src\\index.js":"1","D:\\sites\\juno\\bootcamp\\projects\\project-6\\src\\Verse.js":"2","D:\\sites\\juno\\bootcamp\\projects\\project-6\\src\\Footer.js":"3","D:\\sites\\juno\\bootcamp\\projects\\project-6\\src\\Header.js":"4","D:\\sites\\juno\\bootcamp\\projects\\project-6\\src\\Search.js":"5","D:\\sites\\juno\\bootcamp\\projects\\project-6\\src\\Compose.js":"6","D:\\sites\\juno\\bootcamp\\projects\\project-6\\src\\Haiku.js":"7","D:\\sites\\juno\\bootcamp\\projects\\project-6\\src\\App.js":"8"},{"size":517,"mtime":1607267629837,"results":"9","hashOfConfig":"10"},{"size":1133,"mtime":1607390005214,"results":"11","hashOfConfig":"10"},{"size":278,"mtime":1607277867008,"results":"12","hashOfConfig":"10"},{"size":541,"mtime":1607277867008,"results":"13","hashOfConfig":"10"},{"size":2481,"mtime":1607447530662,"results":"14","hashOfConfig":"10"},{"size":7070,"mtime":1607448273628,"results":"15","hashOfConfig":"10"},{"size":473,"mtime":1607449210029,"results":"16","hashOfConfig":"10"},{"size":1754,"mtime":1607449207720,"results":"17","hashOfConfig":"10"},{"filePath":"18","messages":"19","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},"1m1lww4",{"filePath":"21","messages":"22","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"23","messages":"24","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"25","messages":"26","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"27","messages":"28","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},{"filePath":"29","messages":"30","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"31","messages":"32","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"33","messages":"34","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"D:\\sites\\juno\\bootcamp\\projects\\project-6\\src\\index.js",[],["35","36"],"D:\\sites\\juno\\bootcamp\\projects\\project-6\\src\\Verse.js",[],"D:\\sites\\juno\\bootcamp\\projects\\project-6\\src\\Footer.js",[],"D:\\sites\\juno\\bootcamp\\projects\\project-6\\src\\Header.js",[],"D:\\sites\\juno\\bootcamp\\projects\\project-6\\src\\Search.js",["37"],"D:\\sites\\juno\\bootcamp\\projects\\project-6\\src\\Compose.js",[],"D:\\sites\\juno\\bootcamp\\projects\\project-6\\src\\Haiku.js",[],"D:\\sites\\juno\\bootcamp\\projects\\project-6\\src\\App.js",[],{"ruleId":"38","replacedBy":"39"},{"ruleId":"40","replacedBy":"41"},{"ruleId":"42","severity":1,"message":"43","line":21,"column":53,"nodeType":"44","messageId":"45","endLine":21,"endColumn":55},"no-native-reassign",["46"],"no-negated-in-lhs",["47"],"array-callback-return","Array.prototype.filter() expects a value to be returned at the end of arrow function.","ArrowFunctionExpression","expectedAtEnd","no-global-assign","no-unsafe-negation"]
